moved extra utility function. deleted extra files
